refactor(live): extract browser workflow layer
